fixed menus and user landing
