+++ app/src/main/java/com/example/checkin/ui/AllLocations.kt
@@ -1,10 +1,15 @@
 package com.example.checkin.ui
 
+import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
 import android.widget.Button
 import android.widget.TextView
+import com.example.checkin.R
 import com.example.checkin.databinding.ActivityAllLocationsBinding
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.database.DataSnapshot
@@ -87,9 +92,30 @@ class AllLocations : AppCompatActivity() {
                 binding.allLocationsLl.addView(noneTv)
             }
         })
+    // end onCreate
+    }
 
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        val inflater: MenuInflater = menuInflater
+        inflater.inflate(R.menu.main_menu,menu)
+        return super.onCreateOptionsMenu(menu)
+    }
 
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
 
+        return when(item.itemId) {
+            R.id.main_menu_home -> {
+                var homeIntent = Intent(this, UserLanding::class.java)
+                startActivity(homeIntent)
+                true
+            }
+            R.id.main_menu_profile -> {
+                var profileIntent = Intent(this, ProfileActivity::class.java)
+                startActivity(profileIntent)
+                true
+            }
+            else -> super.onOptionsItemSelected(item)
+        }
 
     }
 }
