+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
              xmlns:app="http://schemas.android.com/apk/res-auto"
-             android:id="@+id/nav_graph">
-
-</navigation>
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph"/>
